Extract config migration methods
